rtor(ECS): Move proper components and systems to server and client dirs
