+++ src/ecs/components/CMakeLists.txt
@@ -2,19 +2,13 @@ set(INCROOT ${PROJECT_SOURCE_DIR}/src/ecs/components)
 set(SRCROOT ${PROJECT_SOURCE_DIR}/src/ecs/components)
 
 set(COMPONENTS_SRC
-    ${SRCROOT}/Controllable.hpp
     ${SRCROOT}/Direction.hpp
-    ${SRCROOT}/Drawable.hpp
-    ${SRCROOT}/EnemyAI.hpp
     ${SRCROOT}/Faction.hpp
     ${SRCROOT}/Health.hpp
     ${SRCROOT}/Position.hpp
-    ${SRCROOT}/Projectile.hpp
-    ${SRCROOT}/Shootable.hpp
     ${SRCROOT}/Size.hpp
     ${SRCROOT}/Velocity.hpp
     ${SRCROOT}/Weapon.hpp
-    ${SRCROOT}/NetworkId.hpp
     ${SRCROOT}/EntityType.hpp
     ${SRCROOT}/EntityLife.hpp
     PARENT_SCOPE
