+++ src/ecs/components/client/CMakeLists.txt
@@ -1,0 +1,14 @@
+set(INCROOT ${PROJECT_SOURCE_DIR}/src/ecs/components/client)
+set(SRCROOT ${PROJECT_SOURCE_DIR}/src/ecs/components/client)
+
+set(COMPONENTS_SRC
+    ${SRCROOT}/Controllable.hpp
+    ${SRCROOT}/Drawable.hpp
+    ${SRCROOT}/Shootable.hpp
+    PARENT_SCOPE
+)
+
+set(COMPONENTS_INCDIRS
+    ${INCROOT}
+    PARENT_SCOPE
+)
