+++ src/ecs/components/client/Drawable.hpp
@@ -16,12 +16,10 @@ namespace ecs::component
      * It differs by constructor parameters
      */
     struct Drawable {
-#ifdef CLIENT_COMPILATION_MODE
         explicit Drawable(const std::filesystem::path &texture) { Texture.loadFromFile(texture.generic_string()); }
         Drawable(const std::filesystem::path &texture, sf::IntRect rect) { Texture.loadFromFile(texture.generic_string(), rect); }
         explicit Drawable(const sf::Texture &texture) : Texture(texture) {}
 
         sf::Texture Texture;
-#endif
     };
 } // namespace ecs::component
