+++ src/ecs/systems/server/DeathUpdate.hpp
